tambah perhitungan

+++ app/Database/Migrations/2021-11-27-145033_Kriteria.php
@@ -18,7 +18,7 @@ class Kriteria extends Migration
                 'constraint'    => '255'
             ],
             'bobot' => [
-                'type'          => 'DOUBLE',
+                'type'          => 'INT',
             ],
             'status'        => [
                 'type'          => 'ENUM',
